src/requirements.txt
anyio==4.9.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
git-filter-repo==2.45.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mplfinance==0.12.10b0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pyparsing==3.2.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-telegram-bot==13.15
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tornado==6.1
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==1.26.16
